acme: directory: make non-required url fields optional
ACME servers provide the directory endpoint for clients to discover
the URLs for ACME operation. These fields are described in RFC 8555
[0], but section 7.1 of the corresponding RFC [1] states that the
server is only required to provide the `directory` and `newNonce`
resources.

Therefore, make the not explicitly required fields optional to not
fail during deserialization of the json response and improve error
handling when the acme client actually requires the not-provided
operation url.

This was discovered based on issues reported in the community forum.

[0] https://www.rfc-editor.org/rfc/rfc8555#section-7.1.1
[1] https://www.rfc-editor.org/rfc/rfc8555#section-7.1

Fixes: https://forum.proxmox.com/threads/159987/
Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250418103849.194119-1-c.ebner@proxmox.com