sys: update nix usage to newer 0.29.1 version from workspace
Mainly adapting to nix adopting more IO Safety by switching the mmap
module and set_tcp_keepalive to accept any AsFd.
Note that the mmap accepting that as value while set_tcp_keepalive
accepting that as reference is mirroring the signatures of the
underlying nix methods for consistency, but nix has no real reason for
this mix either I think.

Besides that most methods that can take a dirfd switched to wrapping
that in an Option, with the main reason being seemingly just because
some other methods already do that, i.e. nix sometimes tries for
consistency...

I also adopted NonNull for the data member in the mmap struct, as nix
switched to that and that type _should_ provide a bit of a safer API.

The newer nix version also updated it's bitflags dependency, which it
uses on lots of places. We mostly notice that in the unsafe
from_bits_unchecked method now being a safe from_bits_retain.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>