http: update to ureq 3
Since ureq 3 is a complete rewrite and switched to using the http
crate, we now depend on http version 1.

We still carry around the 0.2 dependency in public interfaces as long
as we use the old hyper code and need to upgrade this crate again to
use *only* http 1.0 in the future.

For now, for the client-trait impl we just convert the http 1.0
Response to a http 0.2 Response.

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>