rest-server: make minimal example compile again
This was quite definitively applied due to debcargo not running the
tests with all features enabled.
While that indeed cannot be done in general due to potential
conflicting features where that itself will result in compiler errors,
it could be enabled as opt-out, or with some settings for feature
combinations to skip, but I digress.

Fixes: 48fc0a2a ("http: use Watcher of GracefulShutdown to avoid races")
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>