sendmail: fix test failures outside CET/CEST timezones
The `Date` header in the generated email includes a timestamp with a
time zone offset (RFC 2822 format). This causes test failures when the
host system's time zone differs from CET/CEST, as the expected and
actual outputs no longer match.

As a workaround, the `Date` header is excluded from comparisons between
the generated and reference emails.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20250526134211.278900-1-l.wagner@proxmox.com