apt: make test repos separator consistent with other repos
PDM did this right from the beginning, the others will change this for
Trixie.

If this should be fully correct it would need to have the suite passed
as parameter, as is we can only return the default for the one for
which the crate release is currently build for, this is a bit of an
API limitation.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>