apt: avoid extra blank lines in sources output
While a small thing, this often points to an issue where newline
addition was handled by a component which cannot do a proper job.

Like here, the newlines where added in a function that serializes a
single repo entry, it simply has no idea in which context and as such
it really cannot decide if a separating newline after this entry is
warranted or not. So rather move that handling to the place where we
serialize whole files, as there we actually got the required context
to handle this correctly.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>