apt: stop-gap update repo output for trixie
This is not really nicely done, but this crate caused me much more
friction and time required to get it bootstrap than the rest combined,
so it really needs to continue for now, even if still being far from
ideal.

I hinted a few things in a bigger FIXME comment in the tests, but
probably not complete and to be fair, it might not be easy to improve
this substantially, but certainly also not impossible; in short, repos
should be much simpler to add centrally, without having to sprinkle
stuff all over the place, the RepoHandle type might be avoidable, and
if then it really should go, and the test system should be simpler to
edit and shouldn't need any change at all when adding a new distro, or
building for a new one, besides maybe a coverage test to ensure that
all existing distro have some tests.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>