pbs-abi-types: do not declare source as native
it actually would be, it's our code through and through, but we used
the revision for meta updates or if we borked a bump in the past, so
lets keep it for now like (almost) all other librust-* source code
packages from this workspace here.
If we switch, we should switch them all at once for consistency.
And FWIW, this is really not final in any direction, we could switch
back and forth as much as we want as long as something gets bumped,
not that we should ;-)

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>