pbs-api-types: change default notification mode to 'notification-system'
The former default was chose avoid breaking changes during a minor
release, but with the upcoming major release we can change the default.

The rationale for the new default is:
  - 'notification-system' was already used as a default for any datastore
    created by the GUI
  - consistency, since on PVE, the default is 'auto' (which does not exist
    on PBS, but is equivalent to 'notification-system' unless an email
    address has been added as well)
  - users are gently nudged to use the new notification stack

The new default affects any datastore/tape backup job that has no
explicit value set. On the PBS side some migration code will be added to
d/postinst to migrate users who relied on the old default to the
appropriate explicit setting.

As mentioned above, the GUI already explicitly set the value to the new
default for a long time. As a result, this mostly affects newly created
datastores/tape jobs from the CLI/API.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20250623141315.288681-2-l.wagner@proxmox.com