login: use 'PMG' as product for 'PMGQUAR' tickets
We derive the product name from the beginning of the ticket normally,
but this does not work with PMGQUAR tickets. Since the cookie name uses
the product name, this results in a `PMGQUARAuthCookie`, but the cookie
still has to be `PMGAuthCookie` to work. To work around that, decouple
the userid range of the ticket from the product range, and modify it so
that PMGQUAR tickets return also PMG for the product. This will result
in the correct `PMGAuthCookie` name.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20250625073844.609286-1-d.csapak@proxmox.com