api-macro: add allof schema to enum
The API macro required the enum variants to either have a oneOf or
ObjectSchema, but did not allow allOf schemas. There's not really a
reason to not allow allOf as well, since they implement
ObjectSchemaType as well and represent an ObjectSchema, just like
oneOf and ObjectSchema do.

This is in preparation for the SDN fabrics, where sections use the
allOf schema to merge general properties with protocol-specific
properties.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>