s3 client: implement methods to operate on s3 objects in bucket
Adds the basic implementation of the client to use s3 object stores
as backend for PBS datastores.

This implements the basic client actions on a bucket and objects
stored within given bucket.

This is not feature complete and intended to be extended on a
per-demand fashion rather than implementing the whole client at once.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250715125332.954494-7-c.ebner@proxmox.com