s3 client: add example usage for basic operations
Add an examples for how to create the client instance using its
configuration options and how to perform some basic api requests on
the S3 endpoint.

Guarded by the cfg attribute for conditional compilation to not fail
if feature "impl" is not set.

Further, excluded via `Cargo.toml` from being executed as test, as
this requires S3 object store to be available and configured.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250715125332.954494-8-c.ebner@proxmox.com