network-types: add api types for ipv4/6
Add two types that transparently wrap the std structs for IPv4 and
IPv6 addresses as well as MacAddress, so they can be used directly in
structs with the API macro. Similar to their CIDR counterparts, they
have a StringSchema and are (de-)serialized via the respective
FromStr and Display implementations.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250716130837.585796-4-g.goller@proxmox.com