fix #6541: openid: add missing connector chains for http-client
With the upgrade to ureq 3 the TLS connectors (native-tls, rustls) now
require a transport (tcp) in the chain before it, otherwise they
panic.

For HTTP Connect proxy support another ConnectProxy connector is
required.
The new chain, based on the DefaultConnector [0] chain in ureq, needs
to have the connectors in the order of:
ConnectProxy -> Tcp -> Tls

[0] https://github.com/algesten/ureq/blob/3.0.11/src/unversioned/transport/mod.rs#L346

Signed-off-by: Mira Limbeck <m.limbeck@proxmox.com>
Link: https://lore.proxmox.com/20250721144832.570040-1-m.limbeck@proxmox.com
 [TL: tiny commit message touch-up]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>