s3 client: merge secrets config with client config
The secrets config was intended to be stored separately with reduced
access permissions, therefore not part of the regular client config.

In Proxmox Backup Server the config needs however to be accessible
also by the less privileged api endpoints, so this requires the same
permissions anyways. Therefore, drop the dedicates secrets config api
type (still not in use by anything else at the moment) and merge the
secret into the regular client config.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>