notify: sendmail: smtp: list email recipients in stable order
The get_recipients function, which is used by the smtp and sendmail
target, is used to look up user's email addresses from the user
configuration while avoiding duplicate recipients at the same time.

Due to the use of a HashSet to avoid duplicate addresses, the order of
the email addresses was not stable. Only a minor cosmetic detail, but
using the same order as defined in the target config is definitely
nicer, which is achieved by using a plain old Vec instead, adding
.contains(...) checks as needed.

Also add some basic rust doc and test while at it.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20250722120305.326096-2-l.wagner@proxmox.com