s3 client: log error response body for invalid request status codes
While the response body is already logged via `tracing` in case of
unexpected status codes, this is not the case for invalid request
response codes.

Log the error response body for these cases as well, as they might
contain further information to help debug issues. For example,
an incorrectly composed delete_objects call might return the
following error in the response body (reformatted from single line
to multi line for better readability in the commit message):

```
<?xml version="1.0" encoding="UTF-8"?>
<Error>
    <Code>MalformedXML</Code>
    <Message>The XML you provided was not well formed or did not validate against our published schema.</Message>
</Error>
```

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250723102601.495806-4-c.ebner@proxmox.com