proxmox-network-api: use ip link for querying interface information
We only obtained information about whether an interface is active or
not when querying the interfaces of a host. This is not sufficient
anymore with the introduction of proxmox-network-interface-pinning. We
need additional information (MAC address, type of interface, altnames)
to pin network interfaces and to detect them in the rest of the PBS
network stack.

Use 'ip link', analogous to Proxmox VE, instead which provides all the
required information. In the future this could be adapted to query the
kernel via netlink directly, avoiding spawning an additional process.
For this reason, the struct does not expose any internals, so we can
easily switch to netlink in a transparent way for all call sites.
Additionally we also add the information about the altnames of an
interface, obtained from ip link, to the Interface struct.

Part of this has been copied over from proxmox-ve-rs, which already
had versions of the IpLink and AltnameMapping struct.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Tested-by: Christian Ebner <c.ebner@proxmox.com>