s3 client: extend client config and options by optional provider quirks
Add optional provider specific quirks as array to the client
configuration and client options, allowing to set a variant of the
client to follow some custom implementation variants.

As first variant, add the quirk to skip setting the `If-None-Match`
header in put object requests, as some providers do not accept and
implement this headers functionality.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250805105120.623089-2-c.ebner@proxmox.com