auth-api: log `authenticate` result not the return value of `start`
log the actual result of the `authenticate` call in order to make
debugging authentication issues easier. previously we logged the
result of the above `pam_sys::wrapped::start` call, which was
typically `SUCCESS` leading to the very confusing error message:

authentication error - SUCCESS (0)

when in reality the authentication failed.

note that pam uses the SUCCESS return code differently depending on
the function that is called. `pam_start` uses it to communicate that a
transaction was successfully started. `pam_authenticate` uses it to
communicate that a user was successfully logged in.

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20250805154820.324167-1-s.sterz@proxmox.com