s3-client: conditionally compile by `api-types` feature
Currently, a build of s3 client pulls in all dependencies,
independent of the feature set.
To allow correctly building the pbs api types, as required for
e.g. wasm builds using pbs-api-types, add the `api-types` feature
and make the `impl` feature dependencies optional.

Also, set the `api-types` feature to be the default.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250806092122.316117-2-c.ebner@proxmox.com