tfa: copy backup state bits into auth state
Since the webauthn has a flag to allow backup eligibility upgrades,
but uses the *old* eligibility bit to validate the backup *state*, we
need to hack this in now.

We always want to allow eligibility upgrades, since the bits for this
only appeared in 2023 and none of our old data stored them.

It is also unclear whether validating them provides any real benefit.
Software keys can just put arbitrary values in these flags, they don't
necessarily tell the truth.

These flags seem to only have any value if we were to also validate
the keys' certificates and limit users to specific manufacturers where
we "trust" these bits to tell the truth. We don't do that, so we don't
care about these bits.

Instead, hack around this to make the webauthn-rs crate happy.

Additionally, return whether we changed the values, to indicate to the
caller that the local credentials should be upgraded.

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>