s3-client: fix non-optional date header for api response parsing
Fixes an implementation oversight for the http date header response
parsing.

This header is most commonly present for all responses, but it is
documented as optional in the AWS reference documentation [0]. While
introduced during development to possibly detect time shifts between
the local s3 client and the api for e.g. last modified time checks,
it is currently not actively used by any codepath, and can therefore
be adapted without further modifications.

[0] https://docs.aws.amazon.com/AmazonS3/latest/API/RESTCommonResponseHeaders.html

Fixes: https://forum.proxmox.com/threads/169395/
Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/all/20250808114117.393536-2-c.ebner@proxmox.com