s3-client: allow setting custom timeout when sending requests
Currently, all types of requests to the S3 API have the same default
timeout of 60 seconds, which was determined empirically.
Different types of requests to the S3 API will however take quite
different time, especially the uploading of objects via a put call
depends on the payload size and bandwidth available.

Therefore, in preparation for being able to set this on a per-request
basis, allow to optionally set the timeout when sending the request.
This now also allows to send the request without timeout.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>